+++ tests/updater_tests.m
@@ -24,6 +24,17 @@ classdef updater_tests < matlab.unittest.TestCase
             obj.verifyClass(v.string,'char')
         end
 
+        function readVersionsFromFile(obj)
+            % All lines in the file should have valid dates
+            pathToFile = fullfile('updater_data','changelog_dates.txt');
+            % How many lines contain data?
+            fileLines=strsplit(fileread(pathToFile),'\n');
+            numDataLines = sum(cellfun(@(x) length(x),fileLines)>0);
+
+            [~,V] = zapit.updater.getVersionFromChangeLog(pathToFile);
+            obj.verifyEqual(length(V),numDataLines)
+        end
+
         function checkDateIsReasonable(obj)
             d = obj.vStruct.date;
             obj.verifyClass(d.year,'double')
@@ -50,17 +61,17 @@ classdef updater_tests < matlab.unittest.TestCase
         function check_versionStringToStruct(obj)
             vString = 'v0.6.1-beta';
             actual = struct('MAJOR',0, 'MINOR',6, 'PATCH', 1, 'preReleaseString', '-beta', 'string', '0.6.1-beta');
-            vStruct = zapit.updater.versionStringToStructure(vString)
+            vStruct = zapit.updater.versionStringToStructure(vString);
             obj.verifyEqual(actual,vStruct);
 
             vString = '1.6.1-alpha';
             actual = struct('MAJOR',1, 'MINOR',6, 'PATCH', 1, 'preReleaseString', '-alpha', 'string', '1.6.1-alpha');
-            vStruct = zapit.updater.versionStringToStructure(vString)
+            vStruct = zapit.updater.versionStringToStructure(vString);
             obj.verifyEqual(actual,vStruct);
 
             vString = '120.6.1';
             actual = struct('MAJOR',120, 'MINOR',6, 'PATCH', 1, 'preReleaseString', '', 'string', '120.6.1');
-            vStruct = zapit.updater.versionStringToStructure(vString)
+            vStruct = zapit.updater.versionStringToStructure(vString);
             obj.verifyEqual(actual,vStruct);
         end
 
@@ -68,55 +79,55 @@ classdef updater_tests < matlab.unittest.TestCase
             reference = struct('MAJOR',0, 'MINOR',6, 'PATCH', 1, 'preReleaseString', '-beta');
 
             test = struct('MAJOR',0, 'MINOR',6, 'PATCH', 1, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,false);
 
             test = struct('MAJOR',0, 'MINOR',6, 'PATCH', 0, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,false);
 
             test = struct('MAJOR',0, 'MINOR',2, 'PATCH', 0, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,false);
 
             test = struct('MAJOR',0, 'MINOR',2, 'PATCH', 0, 'preReleaseString', '');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,false);
 
             test = struct('MAJOR',0, 'MINOR',2, 'PATCH', 10, 'preReleaseString', '');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,false);
 
             test = struct('MAJOR',0, 'MINOR',6, 'PATCH', 2, 'preReleaseString', '');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',0, 'MINOR',6, 'PATCH', 2, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',0, 'MINOR',7, 'PATCH', 1, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',0, 'MINOR',7, 'PATCH', 0, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',0, 'MINOR',7, 'PATCH', 10, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',1, 'MINOR',0, 'PATCH', 0, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',1, 'MINOR',5, 'PATCH', 0, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
 
             test = struct('MAJOR',1, 'MINOR',5, 'PATCH', 10, 'preReleaseString', '-beta');
-            isNewer = zapit.updater.isVersionNewer(reference,test)
+            isNewer = zapit.updater.isVersionNewer(reference,test);
             obj.verifyEqual(isNewer,true);
         end