+++ zapit/+zapit/+interfaces/@TCPserver/TCPserver.m
@@ -1,23 +1,32 @@
-classdef TCPserver < zapit.interfaces.tcpip
+classdef TCPserver < handle
     % zapit.interfaces.TCPserver
     %
     % This class handles TCP/IP comms. If the tcpServer.enable setting is
     % set to "true" then a server is set up and attached to Zapit using the
-    % port and IP address defined in the settings file. 
-    % 
-    %  
+    % port and IP address defined in the settings file.
+    %
+    %
     properties (Hidden)
-        parent % instance of zaptit.pointer to which we attached
+        parent % instance of zapit.pointer to which we attached
         listeners  % Structure that holds listeners so they can be easily cleaned up in the destructor
-
     end % properties
+
+    properties
+        hSocket % The server or client object will reside here
+        port = 1488
+        ip = 'localhost'
+    end
+
+    properties (SetObservable)
+        buffer
+    end
 
 
     methods
 
         function obj = TCPserver(varargin)
             % Inputs (optional param/val pairs)
-            % 'ip' - [string] Is 'localhost' by default (see zapit.interfaces.tcpip)
+            % 'ip' - [string] Is 'localhost' by default
             % 'port' - [numeric scalar] is 1488 by default
 
             params = inputParser;
@@ -31,17 +40,29 @@ classdef TCPserver < zapit.interfaces.tcpip
             obj.ip = params.Results.ip;
             obj.port = params.Results.port;
 
-             %  ;%localhost';
-
             obj.hSocket = tcpserver(obj.ip, obj.port);
             obj.setupSocket;
 
-            fprintf('Set up TCP server on %s port %d\n', obj.ip, obj.port)
+            fprintf('Set up Zapit TCP server on %s port %d\n', obj.ip, obj.port)
 
             obj.listeners.bufferUpdated = addlistener(obj, 'buffer', 'PostSet', ...
                 @obj.processBufferMessageCallback);
 
         end % Constructor
+
+        function setupSocket(obj)
+            % Set up for reading messages of 4 bytes
+            configureCallback(obj.hSocket,"byte",4,@obj.readDataFcn);
+        end % setupSocket
+
+        function readDataFcn(obj, src, ~)
+            msg = read(src,4,"uint8");
+            obj.buffer = struct('command', msg(1), ...
+                                'ArgKeys', msg(2), ...
+                                'ArgVals', msg(3), ...
+                                'ConditionNumber', msg(4));
+
+        end % readDataFcn
 
 
         function delete(obj)
@@ -50,6 +71,41 @@ classdef TCPserver < zapit.interfaces.tcpip
             delete(obj.hSocket)
         end % Destructor
 
+
+        function varargout = isClientConnected(obj)
+            % Check if a client is currently connected to the server
+            %
+            % function zapit.interfaces.TCPserver.isClientConnected
+            %
+            % Purpose
+            % Return true if a client is connected. False otherwise. If no output
+            % argument is requested, the state is printed to screen.
+            %
+            % Inputs
+            % none
+            %
+            % Outputs
+            % isConnected - [optional] True if connected. False otherwise
+
+            if isempty(obj.hSocket)
+                isConnected = false;
+            else
+                isConnected = obj.hSocket.Connected;
+            end
+
+            if nargout>0
+                varargout{1} = isConnected;
+                return
+            end
+
+            if isConnected
+                fprintf('Client is connected to the TCP server\n')
+            else
+                fprintf('No client is connected to the TCP server\n')
+            end
+        end % isClientConnected
+
     end % methods
 
 end  % TCPserver
+