+++ zapit/+zapit/+updater/getGitInfo.m
@@ -5,28 +5,26 @@ function gitInfo = getGitInfo
 %
 % Purpose
 % Get information about the Git repository in the current directory. Including:
-%          - branch name of the current Git Repo 
-%          -Git SHA1 HASH of the most recent commit
-%          -url of corresponding remote repository, if one exists
+%          - Branch name of the current Git Repo
+%          - Git SHA1 HASH of the mosT commit
+%          - URL of corresponding remote repository, if one exists
 %
 % The function first checks to see if a .git/ directory is present. If so it
 % reads the .git/HEAD file to identify the branch name and then it looks up
 % the corresponding commit.
 %
-% It then reads the .git/config file to find out the url of the
+% It then reads the .git/config file to find out the URL of the
 % corresponding remote repository. This is all stored in a gitInfo struct.
 %
-% Note this uses only file information, it makes no external program 
-% calls at all. 
+% Note this uses only file information, it makes no external program calls at all.
 %
 % This function must be in the base directory of the git repository
 %
-% Released under a BSD open source license. Based on a concept by Marc
-% Gershow.
+% Released a BSD open source license. Based on a concept by Marc Gershow.
 %
 % Andrew Leifer
 % Harvard University
-% Program in Biophysics, Center for Brain Science, 
+% Program in Biophysics, Center for Brain Science,
 % and Department of Physics
 % leifer@fas.harvard.edu
 % http://www.andrewleifer.com
@@ -47,11 +45,11 @@ function gitInfo = getGitInfo
 %
 %    1. Redistributions of source code must retain the above copyright notice, this list of
 %       conditions and the following disclaimer.
-% 
+%
 %    2. Redistributions in binary form must reproduce the above copyright notice, this list
 %       of conditions and the following disclaimer in the documentation and/or other materials
 %       provided with the distribution.
-% 
+%
 % THIS SOFTWARE IS PROVIDED BY <COPYRIGHT HOLDER> ''AS IS'' AND ANY EXPRESS OR IMPLIED
 % WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
 % FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> OR
@@ -61,7 +59,7 @@ function gitInfo = getGitInfo
 % ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 % NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 % ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-% 
+%
 % The views and conclusions contained in the software and documentation are those of the
 % authors and should not be interpreted as representing official policies, either expressed
 % or implied, of the copyright holder.
@@ -179,11 +177,11 @@ for k=1:length(lines)
             temp=textscan(lines{m},'%s');
             if length(temp{1})>=3
                 if strcmp(temp{1}{1},'remote') && strcmp(temp{1}{2},'=')
-                    %This is the line that tells us the name of the remote 
+                    %This is the line that tells us the name of the remote
                     remote=temp{1}{3};
                 end
             end
-            
+
             m=m+1;
         end
     end
@@ -206,11 +204,11 @@ for k=1:length(lines)
             temp=textscan(lines{m},'%s');
             if length(temp{1})>=3
                 if strcmp(temp{1}{1},'url') && strcmp(temp{1}{2},'=')
-                    %This is the line that tells us the name of the remote 
+                    %This is the line that tells us the name of the remote
                     url=temp{1}{3};
                 end
             end
-            
+
             m=m+1;
         end
     end